% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlp.R
\name{activation_h2o}
\alias{activation_h2o}
\title{Parameter function to extend the activation function values for the H2O
deeplearning algorithm}
\usage{
activation_h2o(values = values_activation_h2o)
}
\arguments{
\item{values}{character vector of supported activation functions listed in
`values_activation_h2o`}
}
\value{
A function with classes "qual_param" and "param"
}
\description{
Parameter function to extend the activation function values for the H2O
deeplearning algorithm
}
\examples{
activation_h2o(values = c("Rectifier"))
}
