% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boost_tree_h2o.R
\name{add_boost_tree_h2o}
\alias{add_boost_tree_h2o}
\title{Wrapper to add the `h2o` engine to the parsnip `boost_tree` model
specification}
\usage{
add_boost_tree_h2o()
}
\description{
Wrapper to add the `h2o` engine to the parsnip `boost_tree` model
specification
}
