% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_reg_h2o.R
\name{add_linear_reg_h2o}
\alias{add_linear_reg_h2o}
\title{Wrapper to add the `h2o` engine to the parsnip `linear_reg` model
specification}
\usage{
add_linear_reg_h2o()
}
\description{
Wrapper to add the `h2o` engine to the parsnip `linear_reg` model
specification
}
