% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic_reg.R
\name{add_logistic_reg_h2o}
\alias{add_logistic_reg_h2o}
\title{Wrapper to add the `h2o` engine to the parsnip `logistic_reg` model
specification}
\usage{
add_logistic_reg_h2o()
}
\description{
Wrapper to add the `h2o` engine to the parsnip `logistic_reg` model
specification
}
