% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlp.R
\name{add_mlp_h2o}
\alias{add_mlp_h2o}
\title{Wrapper to add the `h2o` engine to the parsnip `mlp` model specification}
\usage{
add_mlp_h2o()
}
\description{
Wrapper to add the `h2o` engine to the parsnip `mlp` model specification
}
