% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multinomial_reg.R
\name{add_multinom_reg_h2o}
\alias{add_multinom_reg_h2o}
\title{Wrapper to add the `h2o` engine to the parsnip `multinom_reg` model
specification}
\usage{
add_multinom_reg_h2o()
}
\description{
Wrapper to add the `h2o` engine to the parsnip `multinom_reg` model
specification
}
