% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naive_Bayes.R
\name{add_naive_Bayes_h2o}
\alias{add_naive_Bayes_h2o}
\title{Wrapper to add the `h2o` engine to the discrim `naiveBayes` model specification}
\usage{
add_naive_Bayes_h2o()
}
\description{
Wrapper to add the `h2o` engine to the discrim `naiveBayes` model specification
}
