% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rand_forest_h2o.R
\name{add_rand_forest_h2o}
\alias{add_rand_forest_h2o}
\title{Wrapper to add the `h2o` engine to the parsnip `rand_forest` model
specification}
\usage{
add_rand_forest_h2o()
}
\description{
Wrapper to add the `h2o` engine to the parsnip `rand_forest` model
specification
}
