% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automl.R
\name{automl}
\alias{automl}
\title{General interface for automl models}
\usage{
automl(mode = "classification")
}
\arguments{
\item{mode}{A single character string for the type of model.}
}
\value{
A model_spec
}
\description{
General interface for automl models
}
