% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control_h2o.R
\name{control_h2o}
\alias{control_h2o}
\title{Control aspects of the grid search process}
\usage{
control_h2o(
  verbose = FALSE,
  save_pred = FALSE,
  save_models = FALSE,
  event_level = "first"
)
}
\arguments{
\item{verbose}{A logical for logging results as they are generated.}

\item{save_pred}{A logical for whether the out-of-sample predictions should be saved
for each model evaluated.}

\item{save_models}{A logical for whether to retain the models associated with the
tuning and resampling iterations within the h2o cluster and append their h2o model
ids to the resamples object as a '.models' column.}
}
\value{
An object of `control_grid` and `control_resamples` class.
}
\description{
`control_h2o` provides a function to set various aspects of the grid search process.
By default during tuning, the resampling predictions are stored within the h2o
cluster. To save memory and space, use 'save_pred = FALSE'.
``
}
\examples{
# to save space in the cluster use these settings (the defaults)
control_h2o(verbose = TRUE, save_pred = FALSE)
}
