% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{convert_h2o_metrics}
\alias{convert_h2o_metrics}
\title{Converts metrics that are applicable to h2o.getGrid into equivalent yardstick
`metric_set` objects.}
\usage{
convert_h2o_metrics(metric)
}
\arguments{
\item{metric}{A character vector of the yardstick metric, one of c("rsq",
"sensitivity", "rmse", "accuracy", "mn_log_loss", "mse")}
}
\value{
A list with `name` and `metric_set`
}
\description{
Converts metrics that are applicable to h2o.getGrid into equivalent yardstick
`metric_set` objects.
}
