% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tune_grid_h2o.R
\name{extract_h2o_models}
\alias{extract_h2o_models}
\title{Extract the h2o model ids for all of the tuning grids and return them as list
of tibbles.}
\usage{
extract_h2o_models(grid_ids, rename_args)
}
\arguments{
\item{grid_ids}{A character vector of the h2o ids for the tuning grids.}

\item{rename_args}{A named character vector used to remap the h2o
hyperparameter names to tidymodels nomenclature. The names of the vector
are the tidymodels nomenclature and the values are the h2o nomenclature,
e.g. c(mtry = "mtries", min_n = "min_rows")}
}
\value{
a list of tibbles
}
\description{
Extract the h2o model ids for all of the tuning grids and return them as list
of tibbles.
}
\keyword{internal}
