% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_terms.R
\name{extract_terms}
\alias{extract_terms}
\title{Extract response and predictor terms from a formula and a data.frame}
\usage{
extract_terms(formula, data)
}
\arguments{
\item{formula}{formula}

\item{data}{data.frame or tibble}
}
\value{
list, X = character vector of predictor names, y = character vector
of response variable
}
\description{
Extract response and predictor terms from a formula and a data.frame
}
\examples{
data(iris)
vars <- extract_terms(Species ~ ., iris)
}
