% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_predicts.R
\name{gbm_multi_predict}
\alias{gbm_multi_predict}
\title{Multi_predict method for gbm}
\usage{
gbm_multi_predict(object, new_data, type, trees)
}
\arguments{
\item{object}{A parsnip object that inherits from `model_fit` and
`_H2OMultinomialModel`.}

\item{new_data}{A data.frame of new data used for the prediction.}

\item{type}{A character string of either "class", "prob" or "numeric"}

\item{trees}{An integer vector with the number of trees for each submodel.}
}
\value{
A [tibble][tibble::tibble-package].
}
\description{
Multi_predict method for gbm
}
