% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_predicts.R
\name{glm_multi_predict}
\alias{glm_multi_predict}
\title{Multi_predict method for glm}
\usage{
glm_multi_predict(object, new_data, type, penalty)
}
\arguments{
\item{object}{A parsnip object that inherits from `model_fit` and
`_H2OMultinomialModel`.}

\item{new_data}{A data.frame of new data used for the prediction.}

\item{type}{A character string of either "class", "prob" or "numeric"}

\item{penalty}{A numeric vector with the penalty values for different submodels.}
}
\value{
A [tibble][tibble::tibble-package].
}
\description{
Multi_predict method for glm
}
