% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automl.R
\name{h2o_automl_train}
\alias{h2o_automl_train}
\title{Wrapper for training a h2o.automl model}
\usage{
h2o_automl_train(formula, data, model_id, ...)
}
\arguments{
\item{formula}{formula}

\item{data}{data.frame of training data}

\item{model_id}{A randomly assigned identifier for the model. Used to refer
to the model within the h2o cluster.}

\item{...}{Other arguments to pass the h2o.automl}
}
\value{
evaluated h2o model call
}
\description{
Wrapper for training a h2o.automl model
}
