% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm_wrapper.R
\name{h2o_glm_train}
\alias{h2o_glm_train}
\title{Wrapper for training a h2o.glm model as part of a parsnip}
\usage{
h2o_glm_train(formula, data, alpha = NULL, lambda = NULL, family = NULL, ...)
}
\arguments{
\item{formula}{formula}

\item{data}{data.frame of training data}

\item{alpha}{numeric, Distribution of regularization between the L1 (Lasso)
and L2 (Ridge) penalties. A value of 1 for alpha represents Lasso
regression, a value of 0 produces Ridge regression.}

\item{lambda}{numeric, regularization strength}

\item{family}{character, one of c("gaussian", "binomial", "quasibinomial",
"ordinal", "multinomial", "poisson", "gamma", "tweedie",
"negativebinomial")}

\item{...}{other arguments passed to the h2o engine.}
}
\value{
evaluated h2o model call
}
\description{
Wrapper for training a h2o.glm model as part of a parsnip
}
