% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h2o_pred_wrapper.R
\name{h2o_pred}
\alias{h2o_pred}
\title{Wrapper for prediction for h2o mlp model}
\usage{
h2o_pred(object, newdata, ...)
}
\arguments{
\item{object}{H2O model}

\item{newdata}{data to predict}

\item{...}{currently unused}
}
\value{
tibble
}
\description{
Wrapper for prediction for h2o mlp model
}
