% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rule_fit.R
\name{h2o_rulefit_train}
\alias{h2o_rulefit_train}
\title{Wrapper for training a h2o.rulefit model as part of a parsnip
`rule_fit` h2o engine}
\usage{
h2o_rulefit_train(
  formula,
  data,
  rule_generation_ntrees = 50,
  max_rule_length = 3,
  lambda = 0,
  ...
)
}
\arguments{
\item{formula}{formula}

\item{data}{data.frame of training data}

\item{rule_generation_ntrees}{integer, the number of trees to build (default
= 50)}

\item{max_rule_length}{integer, the maximum tree depth (default = 3).}

\item{lambda}{Specify the regularization strength for LASSO regressor.}

\item{...}{other arguments that are passed to the h2o model}
}
\value{
evaluated h2o model call
}
\description{
Wrapper for training a h2o.rulefit model as part of a parsnip
`rule_fit` h2o engine
}
