% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlp.R
\name{h2o_train}
\alias{h2o_train}
\title{Parsnip model specification}
\usage{
h2o_train(
  formula,
  data,
  l2 = 0,
  hidden_dropout_ratios = 0,
  hidden = 200,
  epochs = 10,
  activation = "Rectifier",
  ...
)
}
\arguments{
\item{formula}{formula}

\item{data}{data.frame of training data}

\item{l2}{numeric, l2 regulation parameter, default = 0}

\item{hidden_dropout_ratios}{dropout ratio for a single hidden layer (default
= 0)}

\item{hidden}{integer, number of neurons in the hidden layer (default = 200)}

\item{epochs}{integer, number of epochs (default = 10)}

\item{activation}{character, activation function. Must be one of: "Tanh",
"TanhWithDropout", "Rectifier", "RectifierWithDropout", "Maxout",
"MaxoutWithDropout". Defaults to "Rectifier. If `hidden_dropout_ratios` > 0
then the equivalent activation function with dropout is used.}

\item{...}{other arguments not currently used}
}
\value{
evaluated h2o model call
}
\description{
Parsnip model specification
}
