% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_h2oframe.R
\name{make_h2o_call}
\alias{make_h2o_call}
\title{Create call to a h2o model by merging arguments}
\usage{
make_h2o_call(.fn, args, others)
}
\arguments{
\item{.fn}{character, name of h2o model e.g. h2o.randomForest}

\item{args}{list of arguments to include within call to model}

\item{others}{list of additional args passed through ...}
}
\value{
h2o trained model
}
\description{
Create call to a h2o model by merging arguments
}
