% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tune_grid_h2o.R
\name{model_spec_to_algorithm}
\alias{model_spec_to_algorithm}
\title{Return the equivalent h2o algorithm name for a parsnip `model_spec` object.}
\usage{
model_spec_to_algorithm(object, model_args)
}
\arguments{
\item{object}{A parsnip `model_spec` object.}

\item{model_args}{A list of model arguments.}
}
\value{
A list with algorithm, the model name and the arguments with the
  family attribute set for specific models (e.g. glm).
}
\description{
Return the equivalent h2o algorithm name for a parsnip `model_spec` object.
}
\keyword{internal}
