% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{mse}
\alias{mse}
\alias{mse_vec}
\alias{mse.data.frame}
\title{Mean squared error}
\usage{
mse(data, ...)

mse_vec(truth, estimate, na_rm = TRUE, ...)

\method{mse}{data.frame}(data, truth, estimate, na_rm = TRUE, ...)
}
\arguments{
\item{data}{A `data.frame` containing the `truth` and `estimate` columns.}

\item{...}{Not currently used.}
}
\value{
A `tibble` with columns `.metric`, `.estimator`, and `.estimate` and 1 row of values.
}
\description{
Calculate the mean squared error. This metric is in squared units of the original data.
}
