% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{mse_vec}
\alias{mse_vec}
\alias{mse}
\alias{mse.data.frame}
\title{Mean squared error}
\usage{
mse_vec(truth, estimate, na_rm = TRUE, ...)

mse(data, ...)

\method{mse}{data.frame}(data, truth, estimate, na_rm = TRUE, ...)
}
\arguments{
\item{truth}{The column identifier for the true class results (that is a
factor). This should be an unquoted column name although this argument is
passed by expression and supports quasiquotation (you can unquote column
names). For _vec() functions, a factor vector.}

\item{estimate}{The column identifier for the predicted class results (that
is also factor). As with truth this can be specified different ways but the
primary method is to use an unquoted variable name. For _vec() functions, a
factor vector.}

\item{na_rm}{A logical value indicating whether NA values should be stripped
before the computation proceeds.}

\item{...}{Not currently used.}

\item{data}{A `data.frame` containing the `truth` and `estimate` columns.}
}
\value{
A `tibble` with columns `.metric`, `.estimator`, and `.estimate` and 1 row of values.
}
\description{
Calculate the mean squared error. This metric is in squared units of the original data.
}
