% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boost_tree.R
\name{multi_predict._H2ORegressionModel}
\alias{multi_predict._H2ORegressionModel}
\title{Multi_predict method for h2o gbm regression models}
\usage{
\method{multi_predict}{`_H2ORegressionModel`}(object, new_data, type = "numeric", trees, ...)
}
\arguments{
\item{object}{A `model_spec` object.}

\item{new_data}{A data.frame of new observations.}

\item{type}{A single character vector, must be "numeric".}

\item{trees}{An integer vector with the number of trees in the ensemble.}

\item{...}{Other arguments currently unused.}
}
\value{

}
\description{
Multi_predict method for h2o gbm regression models
}
