% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tune_grid_h2o.R
\name{remove_h2o_models}
\alias{remove_h2o_models}
\title{Removes the models creating during tuning resampling}
\usage{
remove_h2o_models(grid_ids)
}
\arguments{
\item{grid_ids}{A character vector of h2o grid ids for each tuning resample.}
}
\value{

}
\description{
Removes the models creating during tuning resampling
}
\keyword{internal}
