% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tune_grid_h2o.R
\name{tune_grid_h2o}
\alias{tune_grid_h2o}
\title{Tune h2o models}
\usage{
tune_grid_h2o(object, preprocessor = NULL, resamples, grid)
}
\arguments{
\item{object}{A parsnip model spec object.}

\item{preprocessor}{A recipe object.}

\item{resamples}{A rset object.}

\item{grid}{A tibble of hyperparameter values.}
}
\value{

}
\description{
Tune h2o models
}
