% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boost_tree.R
\name{h2o_gbm_train}
\alias{h2o_gbm_train}
\title{Wrapper for training a h2o.gbm model as part of a parsnip `boost_tree`
h2o engine}
\usage{
h2o_gbm_train(
  formula,
  data,
  ntrees = 50,
  max_depth = 5,
  min_rows = 10,
  learn_rate = 0.1,
  sample_rate = 1,
  col_sample_rate = 1,
  min_split_improvement = 1e-05,
  stopping_rounds = 0,
  validation = 0,
  ...
)
}
\arguments{
\item{formula}{formula}

\item{data}{data.frame of training data}

\item{ntrees}{integer, the number of trees to build (default = 50).}

\item{max_depth}{integer, the maximum tree depth (default = 10).}

\item{min_rows}{integer, the minimum number of observations for a leaf
(default = 10).}

\item{learn_rate}{numeric, the learning rate (default = 0.1, range is from
0.0 to 1.0).}

\item{sample_rate}{numeric, the proportion of samples to use to build each
tree (default = 1.0).}

\item{col_sample_rate}{numeric, the proportion of features available during
each node split (default = 1.0).}

\item{min_split_improvement}{numeric,  minimum relative improvement in
squared error reduction in order for a split to happen (default = 1e-05)}

\item{stopping_rounds}{An integer specifying the number of training
iterations without improvement before stopping. If `stopping_rounds = 0`
(the default) then early stopping is disabled.  If `validation` is used,
performance is base on the validation set; otherwise the training set is
used.}

\item{validation}{A positive number. If on `[0, 1)` the value, `validation`
is a random proportion of data in `x` and `y` that are used for performance
assessment and potential early stopping. If 1 or greater, it is the _number_
of training set samples use for these purposes.}

\item{...}{other arguments not currently used.}
}
\value{
evaluated h2o model call
}
\description{
Wrapper for training a h2o.gbm model as part of a parsnip `boost_tree`
h2o engine
}
