% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naive_Bayes.R
\name{h2o_naiveBayes_train}
\alias{h2o_naiveBayes_train}
\title{Wrapper for training a h2o.naiveBayes model as part of a discrim `naive_Bayes`
h2o engine}
\usage{
h2o_naiveBayes_train(formula, data, laplace = 0, ...)
}
\arguments{
\item{formula}{formula}

\item{data}{data.frame of training data}

\item{laplace}{numeric, the Laplace smoothing parameter, must be >= 0.}

\item{...}{other arguments passed to the h2o engine.}
}
\value{
a fitted h2o model.
}
\description{
Wrapper for training a h2o.naiveBayes model as part of a discrim `naive_Bayes`
h2o engine
}
