% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rand_forest.R
\name{h2o_rf_train}
\alias{h2o_rf_train}
\title{Wrapper for training a h2o.randomForest model as part of a parsnip
`rand_forest` h2o engine}
\usage{
h2o_rf_train(
  formula,
  data,
  model_id,
  ntrees = 50,
  min_rows = 10,
  mtries = -1,
  ...
)
}
\arguments{
\item{formula}{formula}

\item{data}{data.frame of training data}

\item{model_id}{A randomly assigned identifier for the model. Used to refer
to the model within the h2o cluster.}

\item{ntrees}{integer, the number of trees to build (default = 50)}

\item{min_rows}{integer, the minimum number of observations for a leaf
(default = 10)}

\item{mtries}{integer, the number of columns to randomly select at each
level. Default of -1 is sqrt(p) for classification and (p/3) for regression.}

\item{...}{other arguments not currently used}
}
\value{
evaluated h2o model call
}
\description{
Wrapper for training a h2o.randomForest model as part of a parsnip
`rand_forest` h2o engine
}
