% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_predicts.R
\name{multi_predict._H2OMultinomialModel}
\alias{multi_predict._H2OMultinomialModel}
\title{Multi_predict method for h2o classification models}
\usage{
\method{multi_predict}{`_H2OMultinomialModel`}(object, new_data, type = "class", ...)
}
\arguments{
\item{object}{A `model_spec` object.}

\item{new_data}{A data.frame of new observations.}

\item{type}{A single character vector of "class" or "prob".}

\item{...}{Other arguments.}
}
\value{

}
\description{
Multi_predict method for h2o classification models
}
