% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tune_grid_h2o.R
\name{tune_grid_h2o}
\alias{tune_grid_h2o}
\title{Tune h2o models}
\usage{
tune_grid_h2o(object, preprocessor = NULL, resamples, grid, verbose = FALSE)
}
\arguments{
\item{object}{A parsnip model spec object.}

\item{preprocessor}{A recipe object.}

\item{resamples}{A rset object.}

\item{grid}{A tibble of hyperparameter values.}

\item{verbose}{Whether to provide progress messages.}
}
\value{

}
\description{
This is a prototype of a version of tune_grid that uses h2o.grid to perform
hyperparameter tuning.
}
\details{
Current limitations
-------------------
- Only model arguments can be tuned
- Parsnip descriptors are not supported
- Custom metrics are not support. Currently r2 is used for regression and
logloss is used for classification.
- Parsnip only allows `data.frame` and `tbl_spark` objects to be passed
to the `fit` method, not `H2OFrame` objects.
}
