% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tune_grid_h2o.R
\name{tune_grid_h2o}
\alias{tune_grid_h2o}
\title{Tune h2o models}
\usage{
tune_grid_h2o(
  object,
  preprocessor = NULL,
  resamples,
  grid,
  metric,
  verbose = FALSE
)
}
\arguments{
\item{object}{A parsnip model spec object.}

\item{preprocessor}{A recipe object.}

\item{resamples}{A rset object.}

\item{grid}{A tibble of hyperparameter values.}

\item{metric}{A `character` vector specifying the h2o metric to use as the
scoring criterion for the hyperparameter tuning. Must be one of  c("rsq",
"sensitivity", "rmse", "accuracy", "mn_log_loss", "mse"). The default is
"rsq" for regression models and "mn_log_loss" for classification models.}

\item{verbose}{Whether to provide progress messages.}
}
\value{

}
\description{
This is a prototype of a version of tune_grid that uses h2o.grid to perform
hyperparameter tuning.
}
\details{
Limitations

- Only model arguments can be tuned, not arguments
in the preprocessing recipes.
- Parsnip only allows `data.frame` and `tbl_spark` objects to be passed
to the `fit` method, not `H2OFrame` objects.
}
