% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tune_grid_h2o.R
\name{tune_grid_h2o}
\alias{tune_grid_h2o}
\title{Tune h2o models}
\usage{
tune_grid_h2o(object, preprocessor = NULL, resamples, grid, metrics)
}
\arguments{
\item{object}{A parsnip model spec object.}

\item{preprocessor}{A recipe object.}

\item{resamples}{A rset object.}

\item{grid}{A tibble of hyperparameter values.}

\item{metrics}{A character vector of metrics to evaluate the tuning results.
Available metrics are r2, mean_per_class_accuracy, max_per_class_error,
err, total_rows, rmse, accuracy, err_count, logloss, mse,
mean_per_class_error}
}
\value{

}
\description{
Tune h2o models
}
